Suppress type conversion during argument pre-parsing
When the 'type' argument is specified when calling
ArgumentParser.add_argument, a type conversion is performed during the
call to ArgumentParser.parse_args. It is normal for this type conversion
to raise an exception if the argument is not well-formed.

This extension first parses the arguments to see which mixins should be
used, and after operating on that information, performs a 'final'
parsing. Since the argument values could change based on the requested
mixins, we should delay raising those exceptions until the 'final'
parsing invocation.