Propagate argument wrapping through default type conversion
If an argument default is a `str` type and the argument is also given an
explicit `type`, the conversion will be performed on the default during
parsing, which strips the wrapper.

This change wraps the type conversion to re-add the wrapper to the
converted value if the original value was wrapped as well.